feat#22: moved Node struct and associated logic to separate file reworked Tidy (now TrimJSON) function - now returns byte slice => fixed all errors added EnsureSaved called to Materialize +minor
