+++ pkg/utils/json-utils.go
@@ -7,13 +7,13 @@ import (
 	"strings"
 )
 
-func Tidy(data string) (string, error) {
+func TrimJSON(data string) ([]byte, error) {
 	start := strings.IndexByte(data, '{')
 	end := strings.LastIndexByte(data, '}')
 	if start == -1 || end == -1 {
-		return "", fmt.Errorf("not a valid json")
+		return nil, fmt.Errorf("not a valid json")
 	}
-	return data[start : end+1], nil
+	return []byte(data)[start : end+1], nil
 }
 
 func SaveJSONToFile(content any, filePath string) error {
